Renamed sequencer.ts -> helpers.ts
